Disable BusyBox watchdog in favor of Troglobit watchdogd
The main purpose of TroglOS is to serve as a test platform for my own
software packages.  This commit introduces my own take on a userspace
watchdog daemon.

The Troglobit watchdogd supports monitoring loadavg, meminfo and filenr
-- all highly useful on an embedded system with no way of adding remote
monitoring of resources.  It also supports process monitoring with an
instrumentation API for client processes.

More to come later, this commit only replaces the BusyBox watchdog.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>