Add Ncurses with -ltermcap compat support
This patch adds ncurses to TroglOS, with -ltermcap support inspired by
Debian.  The libtinfo.{a, so} that ncurses provides is actually fully
compatible with libtermcap, which many older programs look for.

This is a completely unoptimized addition of ncurses, it could be set
up so that we install libtinfo/libtermcap and/or libcurses/libncurses,
depending on user's menuconfig choices.  For the time being all the
libraries are installed.

Header files are installed to staging/include/, not any ncurses sub-
directory, which is the default.  Most programs expect term.h, et al
to be in /usr/include, not in /usr/include/ncurses.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>