Migrate to crosstool-NG based toolchain
This patch cleans up the build system and migrates away from the
Debian/Ubuntu ARM cross toolchain based on GLIBC to a crosstool-NG¹
based ARM toolchain based on musl libc.

By moving to a crosstool-NG toolchain we also get the possiblity to
run the populate script to install only the required toolchain libs
in the initramfs root file system.

____
¹ https://github.com/crosstool-ng/crosstool-ng

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>