User OverlayFS for /var and /etc on /mnt
This patch make /var and /etc writable by setting up an overlay from
/mnt/var and /mnt/etc.

To make sure this works properly the MTD plugin to Finit now creates
the following dirs at boot, if mtd:Config is mounted:

- /mnt/etc
- /mnt/var
- /mnt/.tmp

Also, with the addition of /var as a writable file system we migrate
to a /run on a tmpfs, like many other distros already use.  Setting
it up is the responsibility of Finit, as long as the base directory
exists in the root fs.

With /var writable we can now also move /http to /var/www which means
TroglOS now almost looks like a regular system.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>