qemu: 9p filesystem passthrough, tap networking
In addition to /mnt, use a 9p filesystem passthrough mount on
/host. By default, it uses the $(PERSISTENT) directory, but this can
be overwritten by setting $QEMU_HOSTFS.

Add a "tap" network mode. No virbr0 interface is required, guest's
interface is exported as qtap on the host. The default is still the
"bridge" mode, but since "tap" places fewer requirements on the host,
maybe the default should be "tap" in the future?