qemu: refactor qemu setup
Homogenize qemu's environment variable names under the `QEMU_`
namespace.

Add a "tap" network mode. No `virbr0` interface is required, the
guest's interface is exported as a stand-alone interface called `qtap`
on the host. The default is still the "bridge" mode.

In addition to the JFFS2 filesystem mounted on `/mnt`, use a 9p
filesystem passthrough mount on `/host`. By default, it uses the
default location of the JFFS2 image (typically `~/TroglOS`).