ntpd.conf: Use built-in 'log' function and adjust default condition
The BusyBox ntpd does by default not have support for syslog when
running in the foreground.  There is a TroglOS patch for this, but
this change is a aimed at testing the Finit 'service log' support.

Also, the default condition had eth0 being up hard-coded, that's not
really correct.  All ntpd needs is a default route out.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>