Add 'image' target, shorthand for romfs + ramdisk
This patch adds an image target to the top-level Makefile.  The intended
purpose, for now, is to serve as a shorthand for 'romfs ramdisk'.  This
means you can now more easily fix a bug in a package and run:

    make packages/watchdogd-install image

It is also the beginnings of arch/product specific support for creating
bootable images, with or without a boot loader.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>