Start daemon in normal runlevel and remove dep. on syslgod
The built-in watchdog in Finit is fully capable of serving our needs
until the system is up.  Otherwise watchdogd will have to wait for
/var/run/finit.sock to become available, stealing CPU resources at
bootstrap, which is unnecessary.

Also, remove dependency on syslogd, if syslogd is restarted or stopped
we do not want watchdogd to be restarted as well.  Set <!> though, since
watchdogd does not handle SIGHUP currently.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>