Allow watchdogd to run in all runlevels so it can handle reboot
This patch enables using watchdogd to handle reboot.  This is to allow a
soft-reboot, using `reboot`, to behave like a cold-reboot (power cycle).

In this mode Finit sends SIGPWR to watchdogd at the end of runlevel 6.
Then watchdogd sets the WDT timeout = 1 sec and stops kicking, allowing
the WDT perform the actual system reboot.

The WDT on most embedded systems has its RESETn connected to the RESETn
of all other circuits on the board, i2c devices, CPU, everything except
the RTC.  In this fashion we can make sure the system behaves exactly
the same regardless of how it booted.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>