qemu: Skip "native" serial console ports for each platform
Instead of using the qemu default serial console, we now standardize on
the virtio-console driver.  This provides us with a unified console on
all architectures when running in Qemu.

It also means no more automatic "clear" of the console by Qemu, so it's
easier to scroll-back in the same terminal window to debug.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>