board/common: match lan* as group ports as well as eth*
For boards like the EspressoBin, where we have a switchcore supported
by Linux switchdev, we'd like to set up the bulk of the ports as LAN
ports.  On the EspressoBin these ports are called lan0, lan1, and lan2
on the Ultra.  All these ports should be switch ports and the wan port
should be isolated from the others and act as our upstream port which
we route to/from.

The /lib/system/if script matches and groups ports using the iproute2
command link set group.  This group is later used when bringing up the
bridge and vlan1 interface in /etc/network/interfaces.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>