configs/cn9130: new defconfig
Add support for the Marvell CN9130-CRB-A, a reference board for the
Octeon TX2 SoC along with a Marvel 88E5393X switch chipset.

The linux_defconfig is optimized for the board, based on the default
Arm64 defconfig, but stripped down and complemetned with criticial
drivers to enable board specific features, e.g. the 6393 and the
pinctrl driver used by the I/O expander.

Like most defconfigs in Buildroot, we use an extlinux.conf to give us a
basic menu in U-Boot, which is easier to work with than boot.scr files.

Note: this requires an update of the Buildroot mv-ddr-marvell package
      to work with ATF v2.5: MV_DDR_MARVELL_VERSION = 7753d4b

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>