configs/macchiatobin: new defconfig
The MacchiatoBIN from SolidRun is (was?) the sort of reference board for
the Marvell Armada 8040 SoC.  This defconfig is a direct import of the
upstream Buildroot, with myLinux specific packages like Finit enabled.

Also, this build has been used to verify the new mv-ddr-marvell version
for ATF v2.5, which the myLinux version of this defconfig uses.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>