Initial support for the Nanopi R2S from FriendlyELEC
Support for this is based on the upstream Buildroot R2S defconfig.
For information about the Nanopi R2S, see the FriendlyELC wiki at:

    https://wiki.friendlyelec.com/wiki/index.php/NanoPi_R2S

A set of wifi drivers, and firmware, have been included in this build
to aid in setting up the R2S as an access point.  If the device can
be detected and is supported, it will be set up as wlan0.

Please note:

  - WAN and LAN ports are named eth0 and eth1, respectively
  - The WAN port runs DHCP client
  - The LAN port is attached to a bridge, br0
  - The bridge runs an NTP/DHCP server 192.168.0.100-250
  - The bridge is the only way to connect to the SSH server
  - The bridge has IP adddress 192.168.0.1/24
  - The system automatically detects a wlan0 interface
  - When hostapd has started it adds wlan0 to the bridge, br0
  - A masquerading firewall is set up by default from br0 to WAN

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>